listening-comp/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
chromadb==0.4.18
boto3==1.34.11
pydantic==2.5.2 